[LFXV2-1460] refactor: move GenericFGAMessage types into pkg/types
Extract GenericFGAMessage, GenericAccessData, GenericDeleteData, and
GenericMemberData from handler_generic.go into a shared pkg/types package.
Update handler_generic.go to import and reference the types directly from
fgatypes.* so consumer services can import the canonical definitions.
